Некоторые корректировки
