+++ headers/Book.h
@@ -1,9 +1,5 @@
 #pragma once
 #include <iostream>
-#include <fstream>
-#include <sstream>
-#include <string>
-#include <vector>
 
 enum TypeBook { textBook, novel, researchPaper };
